[glue] Try to detect audioflingerglue library path. JB#40317
Instead of using hardcoded library path defined in hybris.c from
audioflingerglue-devel, look for the library from common locations as
default. For other cases the library path can be given with new module
argument "lib".

Also, if the library is found but cannot be loaded for some reason, fail
gracefully instead of segfaulting.